requirements.txt
streamlit
pandas
plotly
openpyxl
gspread
google-auth